Modularise test fixtures
This is so it is easier to see what fixtures belong to which tests.
