+++ tests/fixtures/fixture_helper_embedding.py
